backend/pyproject.toml
[tool.poetry]
name = "shopify-ai-image-processor"
version = "0.1.0"
description = "AI-powered image processing for Shopify stores"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
python-multipart = "^0.0.6"
celery = "^5.3.6"
redis = "^5.0.1"
supabase = "^1.2.0"
pillow = "^10.2.0"
rembg = "^2.0.50"
numpy = "1.24.3"
python-dotenv = "^1.0.0"
prometheus-client = "^0.19.0"
flower = "^2.0.1"
httpx = "^0.24.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = "^2.6.1"
pydantic-settings = "^2.1.0"
ShopifyAPI = "^12.4.0"
email-validator = "^2.1.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.27"}
asyncpg = "^0.29.0"
onnxruntime = "^1.17.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3 